+++ R/actor_search.R
@@ -2,9 +2,11 @@
 #'
 #' @param query character. search query, Lucene query syntax is recommended when `typeahead = FALSE`.
 #' @param typeahead logical. Use typeahead for search? Default is `FALSE`.
+#' @param limit `r template_var_limit(100)`
 #' @param user `r template_var_user()`
 #' @param pass `r template_var_pass()`
 #' @param auth `r template_var_auth()`
+#' @param clean `r template_var_clean()`
 #'
 #' @concept actor
 #'
@@ -15,14 +17,23 @@
 #' @section Function introduced:
 #' `v0.0.1` (2023-10-01)
 #'
-#' @return a tibble of suggested accounts to follow
+#' @return a [tibble::tibble] of suggested accounts to follow
 #' @export
 #'
 #' @examplesIf has_bluesky_pass() & has_bluesky_user()
 #' bs_search_actors('political science')
-bs_search_actors <- function(query, typeahead = FALSE,
+bs_search_actors <- function(query, typeahead = FALSE, limit = NULL,
                              user = get_bluesky_user(), pass = get_bluesky_pass(),
-                             auth = bs_auth(user, pass)) {
+                             auth = bs_auth(user, pass), clean = TRUE) {
+
+  if (!is.null(limit)) {
+    if (!is.numeric(limit)) {
+      cli::cli_abort('{.arg limit} must be numeric.')
+    }
+    limit <- as.integer(limit)
+    limit <- max(limit, 1L)
+    limit <- min(limit, 100L)
+  }
 
   base_url <- ifelse(
     typeahead,
@@ -31,11 +42,16 @@ bs_search_actors <- function(query, typeahead = FALSE,
   )
   req <- httr2::request(base_url) |>
     httr2::req_url_query(q = query) |>
-    httr2::req_auth_bearer_token(token = auth$accessJwt)
+    httr2::req_auth_bearer_token(token = auth$accessJwt) |>
+    httr2::req_url_query(
+      limit = limit
+    )
 
   resp <- req |>
     httr2::req_perform() |>
     httr2::resp_body_json()
+
+  if (!clean) return(resp)
 
   resp |>
     purrr::pluck('actors') |>