+++ R/feed_search.R
@@ -6,6 +6,7 @@
 # #' @param user `r template_var_user()`
 # #' @param pass `r template_var_pass()`
 # #' @param auth `r template_var_auth()`
+# #' @param clean `r template_var_clean()`
 # #'
 # #' @concept feed
 # #'
@@ -15,14 +16,14 @@
 # #' @section Function introduced:
 # #' `v0.0.1` (2023-10-02)
 # #'
-# #' @return a tibble of suggested accounts to follow
+# #' @return a [tibble::tibble] of suggested accounts to follow
 # #' @export
 # #'
 # #' @examplesIf has_bluesky_pass() & has_bluesky_user()
 # #' bs_search_posts('redistricting')
 # bs_search_posts <- function(query,
 #                              user = get_bluesky_user(), pass = get_bluesky_pass(),
-#                              auth = bs_auth(user, pass)) {
+#                              auth = bs_auth(user, pass), clean = TRUE) {
 #
 #   req <- httr2::request('https://bsky.social/xrpc/app.bsky.feed.searchPosts') |>
 #     httr2::req_url_query(q = query) |>
@@ -31,6 +32,8 @@
 #   resp <- req |>
 #     httr2::req_perform() |>
 #     httr2::resp_body_json()
+#
+#   if (!clean) return(resp)
 #
 #   resp |>
 #     purrr::pluck('posts') |>