+++ R/feed_suggestions.R
@@ -1,8 +1,10 @@
 #' Get a list of feed suggestions
 #'
+#' @param limit `r template_var_limit(100)`
 #' @param user `r template_var_user()`
 #' @param pass `r template_var_pass()`
 #' @param auth `r template_var_auth()`
+#' @param clean `r template_var_clean()`
 #'
 #' @concept feed
 #'
@@ -12,20 +14,35 @@
 #' @section Function introduced:
 #' `v0.0.1` (2023-10-02)
 #'
-#' @return a tibble of suggested feeds
+#' @return a [tibble::tibble] of suggested feeds
 #' @export
 #'
 #' @examplesIf has_bluesky_pass() & has_bluesky_user()
 #' bs_get_feed_suggestions()
-bs_get_feed_suggestions <- function(user = get_bluesky_user(), pass = get_bluesky_pass(),
-                                     auth = bs_auth(user, pass)) {
+bs_get_feed_suggestions <- function(limit = NULL,
+                                    user = get_bluesky_user(), pass = get_bluesky_pass(),
+                                    auth = bs_auth(user, pass), clean = TRUE) {
+
+  if (!is.null(limit)) {
+    if (!is.numeric(limit)) {
+      cli::cli_abort('{.arg limit} must be numeric.')
+    }
+    limit <- as.integer(limit)
+    limit <- max(limit, 1L)
+    limit <- min(limit, 100L)
+  }
 
   req <- httr2::request('https://bsky.social/xrpc/app.bsky.feed.getSuggestedFeeds') |>
-    httr2::req_auth_bearer_token(token = auth$accessJwt)
+    httr2::req_auth_bearer_token(token = auth$accessJwt) |>
+    httr2::req_url_query(
+      limit = limit
+    )
 
   resp <- req |>
     httr2::req_perform() |>
     httr2::resp_body_json()
+
+  if (!clean) return(resp)
 
   resp |>
     purrr::pluck('feeds') |>