+++ R/graph_actor_lists.R
@@ -1,13 +1,15 @@
 #' Get a list of lists that belong to an actor.
 #'
 #' @param actor `r template_var_actor()`
+#' @param limit `r template_var_limit(100)`
 #' @param user `r template_var_user()`
 #' @param pass `r template_var_pass()`
 #' @param auth `r template_var_auth()`
+#' @param clean `r template_var_clean()`
 #'
 #' @concept graph
 #'
-#' @return a tibble of lists
+#' @return a [tibble::tibble] of lists
 #' @export
 #'
 #' @section Lexicon references:
@@ -18,9 +20,9 @@
 #'
 #' @examplesIf has_bluesky_pass() && has_bluesky_user()
 #' bs_get_actor_lists('chriskenny.bsky.social')
-bs_get_actor_lists <- function(actor,
-                           user = get_bluesky_user(), pass = get_bluesky_pass(),
-                           auth = bs_auth(user, pass)) {
+bs_get_actor_lists <- function(actor, limit = NULL,
+                               user = get_bluesky_user(), pass = get_bluesky_pass(),
+                               auth = bs_auth(user, pass), clean = TRUE) {
 
   if (missing(actor)) {
     cli::cli_abort('{.arg actor} must list at least one user.')
@@ -28,16 +30,29 @@ bs_get_actor_lists <- function(actor,
   if (!is.character(actor)) {
     cli::cli_abort('{.arg actor} must be a character vector.')
   }
+  if (!is.null(limit)) {
+    if (!is.numeric(limit)) {
+      cli::cli_abort('{.arg limit} must be numeric.')
+    }
+    limit <- as.integer(limit)
+    limit <- max(limit, 1L)
+    limit <- min(limit, 100L)
+  }
 
   req <- httr2::request('https://bsky.social/xrpc/app.bsky.graph.getLists') |>
     httr2::req_url_query(actor = actor) |>
-    httr2::req_auth_bearer_token(token = auth$accessJwt)
+    httr2::req_auth_bearer_token(token = auth$accessJwt) |>
+    httr2::req_url_query(
+      limit = limit
+    )
   resp <- req |>
     httr2::req_perform() |>
     httr2::resp_body_json()
 
-    resp |>
-      purrr::pluck('lists') |>
-      proc() |>
-      clean_names()
+  if (!clean) return(resp)
+
+  resp |>
+    purrr::pluck('lists') |>
+    proc() |>
+    clean_names()
 }