+++ R/graph_mutes.R
@@ -1,12 +1,14 @@
 #' Retrieve a user's (self) muted accounts
 #'
+#' @param limit `r template_var_limit(100)`
 #' @param user `r template_var_user()`
 #' @param pass `r template_var_pass()`
 #' @param auth `r template_var_auth()`
+#' @param clean `r template_var_clean()`
 #'
 #' @concept graph
 #'
-#' @return a tibble of actors
+#' @return a [tibble::tibble] of actors
 #' @export
 #'
 #' @section Lexicon references:
@@ -17,14 +19,29 @@
 #'
 #' @examplesIf has_bluesky_pass() && has_bluesky_user()
 #' bs_get_mutes()
-bs_get_mutes <- function(user = get_bluesky_user(), pass = get_bluesky_pass(),
-                         auth = bs_auth(user, pass)) {
+bs_get_mutes <- function(limit = NULL,
+    user = get_bluesky_user(), pass = get_bluesky_pass(),
+                         auth = bs_auth(user, pass), clean = TRUE) {
+
+  if (!is.null(limit)) {
+    if (!is.numeric(limit)) {
+      cli::cli_abort('{.arg limit} must be numeric.')
+    }
+    limit <- as.integer(limit)
+    limit <- max(limit, 1L)
+    limit <- min(limit, 100L)
+  }
 
   req <- httr2::request('https://bsky.social/xrpc/app.bsky.graph.getMutes') |>
-    httr2::req_auth_bearer_token(token = auth$accessJwt)
+    httr2::req_auth_bearer_token(token = auth$accessJwt) |>
+    httr2::req_url_query(
+      limit = limit
+    )
   resp <- req |>
     httr2::req_perform() |>
     httr2::resp_body_json()
+
+  if (!clean) return(resp)
 
   resp |>
     purrr::pluck('mutes') |>